+++ addon/globalPlugins/instantTimer/paths.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Tymer Add-on for NVDA
+# Instant Timer Add-on for NVDA
 # Author: Kamal Yaser <kamalyaser31@gmail.com>
 
 import os
@@ -7,7 +7,7 @@ from typing import Final
 import addonHandler
 
 PLUGIN_DIR: Final = os.path.join(
-    addonHandler.getCodeAddon().path, "globalPlugins", "tymer"
+    addonHandler.getCodeAddon().path, "globalPlugins", "instantTimer"
 )
 SOUNDS_DIR: Final = os.path.join(PLUGIN_DIR, "waves")
 ALARM_SOUND_PATH: Final = os.path.join(SOUNDS_DIR, "alarm.wav")